refactor code and finish Day 5
