+++ src/main/kotlin/CreatureList.kt
@@ -1,0 +1,4 @@
+// Class for a list of creatures
+data class CreatureList(
+    val data: MutableList<Creature>
+)
