+++ src/main/kotlin/Planet.kt
@@ -1,0 +1,4 @@
+// Enumeration of Planet
+enum class Planet {
+    Earth, Asgard, Betelgeuse, Vogsphere, Kashyyyk, Endor
+}
