+++ src/main/kotlin/Traits.kt
@@ -1,0 +1,4 @@
+// Enumeration of Traits
+enum class Traits {
+    HAIRY, TALL, SHORT, BLONDE, EXTRA_ARMS, EXTRA_HEAD, GREEN, BULKY, POINTY_EARS
+}
