+++ src/main/kotlin/Universe.kt
@@ -1,0 +1,5 @@
+// Class for defining universe
+data class Universe(
+    val name: String,
+    val individuals: MutableList<Creature> = mutableListOf()
+)
